+++ docker/entrypoint.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+caddy start --config /etc/caddy/Caddyfile --adapter caddyfile
+
+exec $@